Tidying before release.

+++ GCodes.cpp
@@ -1632,12 +1632,12 @@ bool GCodes::HandleMcode(int code, GCodeBuffer *gb)
 
 		if(!seen)
 		{
-			snprintf(reply, STRING_LENGTH, "Steps/mm: X: %d, Y: %d, Z: %d, E: ",
-					(int)platform->DriveStepsPerUnit(X_AXIS), (int)platform->DriveStepsPerUnit(Y_AXIS),
-					(int)platform->DriveStepsPerUnit(Z_AXIS));
+			snprintf(reply, STRING_LENGTH, "Steps/mm: X: %.4f, Y: %.4f, Z: %.4f, E: ",
+					platform->DriveStepsPerUnit(X_AXIS), platform->DriveStepsPerUnit(Y_AXIS),
+					platform->DriveStepsPerUnit(Z_AXIS));
 			for(int8_t drive = AXES; drive < DRIVES; drive++)
 			{
-				snprintf(scratchString, STRING_LENGTH, "%f", platform->DriveStepsPerUnit(drive));
+				snprintf(scratchString, STRING_LENGTH, "%.4f", platform->DriveStepsPerUnit(drive));
 				strncat(reply, scratchString, STRING_LENGTH);
 				if(drive < DRIVES-1)
 					strncat(reply, ":", STRING_LENGTH);
